main 추가/로그인 페이지 파일명 변경(activity_login, LoginActivity)
