+++ app/src/main/res/drawable/ic_dot.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="17dp"
+    android:height="5dp"
+    android:viewportWidth="17"
+    android:viewportHeight="5">
+  <path
+      android:pathData="M0.6162,2.2148C0.6162,1.7051 0.7773,1.2861 1.0996,0.958C1.4277,0.6299 1.8613,0.4658 2.4004,0.4658C2.9336,0.4658 3.3672,0.627 3.7012,0.9492C4.0352,1.2656 4.2021,1.6963 4.2021,2.2412V2.5664C4.2021,3.0762 4.041,3.4922 3.7188,3.8145C3.3965,4.1367 2.96,4.2979 2.4092,4.2979C1.8643,4.2979 1.4277,4.1367 1.0996,3.8145C0.7773,3.4863 0.6162,3.0645 0.6162,2.5488V2.2148ZM6.6807,2.2148C6.6807,1.7051 6.8418,1.2861 7.1641,0.958C7.4922,0.6299 7.9258,0.4658 8.4648,0.4658C8.998,0.4658 9.4316,0.627 9.7656,0.9492C10.0996,1.2656 10.2666,1.6963 10.2666,2.2412V2.5664C10.2666,3.0762 10.1055,3.4922 9.7832,3.8145C9.4609,4.1367 9.0244,4.2979 8.4736,4.2979C7.9287,4.2979 7.4922,4.1367 7.1641,3.8145C6.8418,3.4863 6.6807,3.0645 6.6807,2.5488V2.2148ZM12.7451,2.2148C12.7451,1.7051 12.9062,1.2861 13.2285,0.958C13.5566,0.6299 13.9902,0.4658 14.5293,0.4658C15.0625,0.4658 15.4961,0.627 15.8301,0.9492C16.1641,1.2656 16.3311,1.6963 16.3311,2.2412V2.5664C16.3311,3.0762 16.1699,3.4922 15.8477,3.8145C15.5254,4.1367 15.0889,4.2979 14.5381,4.2979C13.9932,4.2979 13.5566,4.1367 13.2285,3.8145C12.9062,3.4863 12.7451,3.0645 12.7451,2.5488V2.2148Z"
+      android:fillColor="#000000"/>
+</vector>
