+++ app/src/main/res/drawable/ic_profile.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="44dp"
+    android:height="44dp"
+    android:viewportWidth="44"
+    android:viewportHeight="44">
+  <path
+      android:pathData="M5,0L39,0A5,5 0,0 1,44 5L44,39A5,5 0,0 1,39 44L5,44A5,5 0,0 1,0 39L0,5A5,5 0,0 1,5 0z"
+      android:fillColor="#C4C4C4"/>
+</vector>
