+++ app/src/main/res/drawable/ic_search.xml
@@ -1,0 +1,16 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="15dp"
+    android:height="14dp"
+    android:viewportWidth="15"
+    android:viewportHeight="14">
+  <path
+      android:strokeWidth="1"
+      android:pathData="M5.6875,5.6875m-5.1875,0a5.1875,5.1875 0,1 1,10.375 0a5.1875,5.1875 0,1 1,-10.375 0"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"/>
+  <path
+      android:strokeWidth="1"
+      android:pathData="M9.9785,9.2715L14.3536,13.6464"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"/>
+</vector>
