+++ app/src/main/res/drawable/ic_search_bar.xml
@@ -1,0 +1,11 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="268dp"
+    android:height="35dp"
+    android:viewportWidth="268"
+    android:viewportHeight="35">
+  <path
+      android:strokeWidth="1"
+      android:pathData="M5,0.5L263,0.5A4.5,4.5 0,0 1,267.5 5L267.5,30A4.5,4.5 0,0 1,263 34.5L5,34.5A4.5,4.5 0,0 1,0.5 30L0.5,5A4.5,4.5 0,0 1,5 0.5z"
+      android:fillColor="#00000000"
+      android:strokeColor="#C4C4C4"/>
+</vector>
