+++ app/src/main/res/menu/popup.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item
+        android:id="@+id/action_menu1"
+        android:title="모임 삭제" />
+    <item
+        android:id="@+id/action_menu2"
+        android:title="모임명 수정" />
+    <item
+        android:id="@+id/action_menu3"
+        android:title="인원 추가" />
+
+</menu>
